Stop ignoring errors when making the initial symlink
This was my cheap and hacky way to not fail when the link already exists
in a desired state; it wasn't perfect because if the symlink fails to
get created we really ought to know that.

I wanted to just catch EEXIST, but I legitimately can't figure out how.
NSError refuses to pattern-match.