Move let next to its binding
This is a point from Google's style guide, focused on the case where
you're matching a case using the value of a variable (and thus _don't_
want to introduce a new binding). That's not an issue for me, but the
consistency of the enum variant appearing right next to the case keyword
is nice too.