Add an initial swift-format configuration
This matches my current Xcode settings: tabs for indentation with line
breaking based on an assumption of 4-space tabs.

Note that my use of tabs for indentation is an _intentional_ break from
typical Swift style, and is done to make this code more accessible to
more readers for whom 2-space indentation may be difficult to read.
The 4-space assumption is sort of a balance between the 2 spaces Swift
tends to want and the 8 spaces that typically make up a tab.

If .swift-format weren't JSON I'd comment this inline, but for now a Git
blame will have to do.