Set lineBreakBeforeEachArgument
This is closer to the style I've been using, but with the difference
that it breaks trailing parentheses onto their own lines. I don't have a
strong preference against that style, especially not in the face of
having this handled automatically.

I've also reformatted several calls to match my new indentation settings
that assume 4-space tabs. Several of them can now be written on one
line.