Integrate Lightswitch into Lighter
It works perfectly! It correctly flips the symlink!

In its sandbox in ~/Library/Containers.

Whoops.

AFAICT the easiest way out of this will be to turn sandboxing off.
I think I could figure out how to use a security-scoped URL bookmark for
persistent access to ~/.lights alone, but I'm not sure if it would get
mad about executing the hooks at that point. Not like I'm trying to get
this onto the Mac App Store anyways.