Replace MIT License with CC0 dedication
In a previous commit, I briefly alluded to my use of an LLM in the
introduction of Lighter.app to the Lights project. However, I have
failed to properly deal with multiple ethical considerations associated
with this.

As a starting point toward resolving this ethical debt, I'm choosing to
replace the project's MIT licensing with a CC0 dedication, effectively
stripping myself (and my heirs and successors) of all potential benefits
afforded by copyright protection, as clearly laid out in the included
CC0 text.

There are at least two factors that make this decision reasonable and
appropriate.

First, while I have made certain changes of my own to the LLM-generated
portions of Lighter.app, and have integrated it with code I can attest
to having written myself, it's reasonable to argue that Lighter.app
could not have existed in anything close to its current form without the
use of non-human authorship. I generally find myself of the opinion that
works whose essential creative elements are not the product of human
authorship are not deserving of copyright protection, and I believe that
until now I've failed to consider the extent to which this is the case
for Lighter.app. Rather than attempt to tease out my own extensions to
the machine output to argue in favor of myself having meaningfully
authored the work, I think the simplest approach is to treat the
entirety of the project moving forward as outside of the scope of
copyright protection.

Second, because it can be reasonably argued that the technology used to
generate a substantial element of this project is itself the product of
widespread intellectual property theft, it seems hypocritical of me to
enjoy the fruits of intellectual property protection for myself (and my
heirs and successors), at minimum with respect to this project. The
dedication of the non-Lighter.app portions of the project under CC0 may
be taken as an "interest payment" toward the ethical debt referenced
above.

The fact that this action does not touch on implications of LLM training
and use beyond the domain of copyright should not be construed as total
ignorance on my part of such implications. (See, for example, the brief
mention of environmental impacts in the commit introducing Lighter.app.)